+++ packages/elk/src/elk/constants/elk.constant.ts
@@ -3,3 +3,15 @@ export const ELK_MODULE_OPTIONS = Symbol('ELK_MODULE_OPTIONS');
 
 export const ELK_MSG_SUCCESSFULLY_CONNECTED = 'Connected successfully!:';
 export const ELK_MSG_ERROR_CONNECTED = 'Could not establish a connection:';
+
+/*
+export const OMITS_IN_OPTIONS: string[] = ['indexDate', 'redact'];
+export const DEFAULT_OPTIONS_REDACT = {
+  paths: [],
+  censor: '****',
+  remove: false,
+};
+
+export const DEFAULT_TIMEZONE = 'America/Argentina/Buenos_Aires';
+export const DEFAULT_TIME_FORMAT = 'D tt'; // 4/10/2023 12:55:40 PM
+*/